Add files via upload
I put the function for adjusting scale ('adjustVerticalScale') in its own file. Also, the pulse amplitude was both an input and output for this function, and this was resulting in an error. I now edited the function so that it does not return an output. Ismail
