+++ Voltage_Pulsed_IV.py
@@ -17,20 +17,13 @@ from Browse_buttons import browse_plot_file, browse_txt_file
 # from Update_Trigger import updateTriggerCursor
 
 class VPulse_IV():
-    def adjustVerticalScale(self, measChannel, triggerChannel, pulseAmplitude, availableDisplay, verticalScale):
-        while (pulseAmplitude > 0.9*availableDisplay):
-            verticalScale = self.incrOscVertScale(verticalScale)
-            availableDisplay = 6*verticalScale
-            self.scope.write(":CHANNEL%d:SCALe %.3f" % (self.channel, float(verticalScale)))
-            pulseAmplitude = self.scope.query_ascii_values("SINGLE;*OPC;:MEASure:VAMPlitude? CHANNEL%d" % self.channel.get())[0]
-            if (measChannel == triggerChannel):
-                self.updateTriggerCursor(pulseAmplitude, self.scope, availableDisplay)
-            return pulseAmplitude
 
     # Import Oscilloscope scaling
     from Oscilloscope_Scaling import incrOscVertScale
     # Import trigger updating
     from Update_Trigger import updateTriggerCursor
+    # Import function for adjusting vertical scales in oscilloscope
+    from adjustVerticalScale import adjustVerticalScale
 
     def start_iv_pulse(self):
         # Range of values for vertical scale on oscilloscope
@@ -141,12 +134,12 @@ class VPulse_IV():
                     self.updateTriggerCursor(voltage_ampl_osc, self.scope, totalDisplayVoltage)
                     
                 # Adjust vertical scales if measured amplitude reaches top of screen (90% of display)
-                current_ampl_osc = self.adjustVerticalScale(self, self.current_channel.get(), self.trigger_channel.get(),\
+                self.adjustVerticalScale(self.current_channel.get(), self.trigger_channel.get(),\
                     current_ampl_osc, totalDisplayCurrent, vertScaleCurrent)
-                voltage_ampl_osc = self.adjustVerticalScale(self, self.voltage_channel.get(), self.trigger_channel.get(),\
+                self.adjustVerticalScale(self.voltage_channel.get(), self.trigger_channel.get(),\
                     voltage_ampl_osc, totalDisplayVoltage, vertScaleVoltage)
 
-                current_ampl_device = 2*current_ampl_osc
+                current_ampl_device = 2*float(current_ampl_osc)
                 voltage_ampl_device = voltage_ampl_osc - seriesResistance*current_ampl_device
 
                 voltageData.append(voltage_ampl_device)
